Reorganise app structure
