+++ app.js
@@ -3,6 +3,8 @@
  *
  * @author Norbert
  */
+
+// Standard Library
 const {join, resolve} = require('path');
 const {mkdirSync, existsSync} = require('fs');
 
@@ -16,7 +18,7 @@ function rootPath(fileName) {
   return resolve(join(__dirname, fileName));
 }
 
-if (!existsSync('logs')) mkdirSync('logs');
+if (!existsSync(rootPath('logs'))) mkdirSync(rootPath('logs'));
 
 const express = require('express');
 const SECRET = 'U\x0bQ*kf\x1bb$Z\x13\x03\x15w\'- f\x0fn1\x0f\\\x106V\'M~\x07';
@@ -24,9 +26,10 @@ const SECRET = 'U\x0bQ*kf\x1bb$Z\x13\x03\x15w\'- f\x0fn1\x0f\\\x106V\'M~\x07';
 const app = express();
 
 app.use((req, res, next) => {
-  res.header('Access-Control-Allow-Origin', '127.0.0.1');
+  // res.header('Access-Control-Allow-Origin', '127.0.0.1');
   res.header('Access-Control-Allow-Credentials', true);
-  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
+  // res.header('Access-Control-Allow-Headers',
+  //   'Origin, X-Requested-With, Content-Type, Accept');
   next();
 });
 
@@ -47,24 +50,14 @@ app.use(require('cors')({
   credentials: true,
 }));
 
-// app.use(require('cookie-session')({
-// name: 'session',
-// secret: SECRET,
-// // 4 hours
-// maxAge: 4 * 60 * 60 * 1000,
-// }));
-
 // view engine setup
 app.set('views', rootPath('views'));
 app.set('view engine', 'hbs');
 
-app.use(require('morgan')('dev'));
+app.use(require('morgan')(':method :url :status :remote-user :req[cookie]'));
+
 app.use(express.json());
-app.use(
-  express.urlencoded({
-    extended: false,
-  }),
-);
+app.use(express.urlencoded({extended: false}));
 
 app.use(
   require('node-sass-middleware')({
@@ -80,9 +73,15 @@ app.use(
 
 app.use(express.static(rootPath('public')));
 
+app.post((req, res, next) => {
+  console.warn(`req.body: ${req.body}`);
+  console.warn(`req.cookies: ${req.cookies}`);
+  next();
+});
+
 app.use('/user', require('./routes/user'));
 app.use('/api', require('./routes/api'));
-app.use('/', require('./routes/index'));
+app.use('/', require('./routes'));
 
 // catch 404 and forward to error handler
 app.use((req, res, next) => next(require('http-errors')(404)));
@@ -92,6 +91,7 @@ app.use((err, req, res, next) => {
   /** @namespace res.locals */
   // set locals, only providing error in development
   res.locals.message = err.message;
+  /** @namespace req.app */
   res.locals.error = req.app.get('env') === 'development' ? err : {};
 
   // render the error page
