+++ routes/index.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const router = express.Router();
 
-router.get('/',
-  (req, res) => res.render('index', {title: 'Final Year CO600 Project'}));
+router.get('/', (req, res) => res.render('index', {title: 'Final Year CO600 Project'}));
 
 module.exports = router;
