--- routes/lib.test.js
@@ -1,5 +1,0 @@
-/**
- * Tests for utility functions in `./lib.js`.
- *
- * @author Norbert
- */
